+++ index.html
@@ -8,6 +8,8 @@
     <title>MK - JSMarathon - Arenas</title>
     <link rel="stylesheet" href="style.css">
     <link rel="stylesheet" href="custom.css">
+    <link rel="icon" type="image/png" sizes="32x32" href="./assets/favicon-32x32.png">
+    <link rel="icon" href="/favicon.ico" type="image/x-icon">
 </head>
 <body>
     <div class="root">
@@ -60,6 +62,6 @@
         <div class="wall-left"></div>
         <div class="wall-right"></div>
     </div>
-    <script src="main.js"></script>
+    <script src="scripts/main.js" type="module"></script>
 </body>
 </html>
