+++ scripts/characters.js
@@ -1,0 +1,98 @@
+export const characters = [
+    {
+        name: 'scorpion',
+        hp: 100,
+        img: 'http://reactmarathon-api.herokuapp.com/assets/scorpion.gif',
+        weapon: [ 'low punch', 'high punch', 'low kick', 'high kick', 'sword', ],
+        attack: function(){
+            console.log(this.name + ' Fight...')
+        },
+    },
+    {
+        name: 'kitana',
+        hp: 100,
+        img: 'http://reactmarathon-api.herokuapp.com/assets/kitana.gif',
+        weapon: [ 'low punch', 'high punch', 'low kick', 'high kick', 'sword', ],
+        attack: function(){
+            console.log(this.name + ' Fight...')
+        },
+    },
+    {
+        name: 'liu kang',
+        hp: 100,
+        img: 'http://reactmarathon-api.herokuapp.com/assets/liukang.gif',
+        weapon: [ 'low punch', 'high punch', 'low kick', 'high kick', 'sword', ],
+        attack: function(){
+            console.log(this.name + ' Fight...')
+        },
+    },
+    {
+        name: 'sonya',
+        hp: 100,
+        img: 'http://reactmarathon-api.herokuapp.com/assets/sonya.gif',
+        weapon: [ 'low punch', 'high punch', 'low kick', 'high kick', 'sword', ],
+        attack: function(){
+            console.log(this.name + ' Fight...')
+        },
+    },
+    {
+        name: 'sub-zero',
+        hp: 100,
+        img: 'http://reactmarathon-api.herokuapp.com/assets/subzero.gif',
+        weapon: [ 'low punch', 'high punch', 'low kick', 'high kick', 'sword', ],
+        attack: function(){
+            console.log(this.name + ' Fight...')
+        },
+    },
+
+]
+
+const scorpion = {
+    name: 'scorpion',
+    hp: 100,
+    img: 'http://reactmarathon-api.herokuapp.com/assets/scorpion.gif',
+    weapon: [ 'low punch', 'high punch', 'low kick', 'high kick', 'sword', ],
+    attack: function(){
+        console.log(this.name + ' Fight...')
+    },
+}
+
+const kitana = {
+    name: 'kitana',
+    hp: 100,
+    img: 'http://reactmarathon-api.herokuapp.com/assets/kitana.gif',
+    weapon: [ 'low punch', 'high punch', 'low kick', 'high kick', 'sword', ],
+    attack: function(){
+        console.log(this.name + ' Fight...')
+    },
+}
+
+const liukang = {
+    name: 'liu kang',
+    hp: 100,
+    img: 'http://reactmarathon-api.herokuapp.com/assets/liukang.gif',
+    weapon: [ 'low punch', 'high punch', 'low kick', 'high kick', 'sword', ],
+    attack: function(){
+        console.log(this.name + ' Fight...')
+    },
+}
+
+const sonya = {
+    name: 'sonya',
+    hp: 100,
+    img: 'http://reactmarathon-api.herokuapp.com/assets/sonya.gif',
+    weapon: [ 'low punch', 'high punch', 'low kick', 'high kick', 'sword', ],
+    attack: function(){
+        console.log(this.name + ' Fight...')
+    },
+}
+
+const subZero = {
+    name: 'sub-zero',
+    hp: 100,
+    img: 'http://reactmarathon-api.herokuapp.com/assets/subzero.gif',
+    weapon: [ 'low punch', 'high punch', 'low kick', 'high kick', 'sword', ],
+    attack: function(){
+        console.log(this.name + ' Fight...')
+    },
+}
